% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgeRAnaFit.R
\name{edgeRAnaFit}
\alias{edgeRAnaFit}
\title{A edgeRAnaFit Function}
\usage{
edgeRAnaFit(files, dir, group, outprefix, spe)
}
\arguments{
\item{files}{a dataframe which contain edgeR pathway analysis results}

\item{dir}{directory which count files are located}

\item{group}{experiment condition, which is a factor}

\item{outprefix}{the prefix of output file}
}
\description{
This function generate the fit model based on edgeR
}
\examples{
edgeRAnaFit(pathway, up=FALSE)
edgeRAnaFit()
}
\keyword{analysis}
\keyword{pathway}
