% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgeRAnaGO.R
\name{edgeRAnaGO}
\alias{edgeRAnaGO}
\title{A edgeRAnaGO Function}
\usage{
edgeRAnaGO(lrts, fdr = 0.05, dir, comparisons, outprefix, spe)
}
\arguments{
\item{lrts}{a list of lrt fit object which generated by edgeRdfGenes}

\item{fdr}{false discovery rate cutoff for differentially expressed genes. Numeric value between 0 and 1}

\item{dir}{directory which count files are located}

\item{comparisons}{pairwise comparisons}

\item{outprefix}{the prefix of output file}
}
\description{
This function did go analysis
}
\examples{
edgeRAnaGO(pathway, up=FALSE)
edgeRAnaGO()
}
\keyword{analysis}
\keyword{pathway}
