% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgeRAnaKEGG.R
\name{edgeRAnaKEGG}
\alias{edgeRAnaKEGG}
\title{A edgeRAnaKEGG Function}
\usage{
edgeRAnaKEGG(lrts, dfCutOff, dir, comparisons, outprefix, spe)
}
\arguments{
\item{lrts}{a list of lrt fit object which generated by edgeRdfGenes}

\item{dfCutOff}{false discovery rate cutoff for differentially expressed genes. Numeric value between 0 and 1}

\item{dir}{directory which count files are located}

\item{comparisons}{pairwise comparisons}

\item{outprefix}{the prefix of output file}
}
\description{
This function did KEGG pathway analysis based on edgeR
}
\examples{
edgeRAnaKEGG(pathway, up=FALSE)
edgeRAnaKEGG()
}
\keyword{analysis}
\keyword{pathway}
