% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgeRAnaMapedGene2Pathway.R
\name{edgeRAnaMapedGene2Pathway}
\alias{edgeRAnaMapedGene2Pathway}
\title{A edgeRAnaMapedGene2Pathway Function}
\usage{
edgeRAnaMapedGene2Pathway(pathways, df, out = T, dir = NA, outprefix = NA,
  spe)
}
\arguments{
\item{pathways}{a list of lrt fit object which generated by edgeRdfGenes}

\item{df}{false discovery rate cutoff for differentially expressed genes. Numeric value between 0 and 1}
}
\description{
This function generate the fit model based on edgeR
}
\examples{
edgeRAnaMapedGene2Pathway(pathways, df)
edgeRAnaMapedGene2Pathway()
}
\keyword{analysis}
\keyword{pathway}
