% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgeRAnaPathview.R
\name{edgeRAnaPathview}
\alias{edgeRAnaPathview}
\title{A edgeRAnaPathview Function}
\usage{
edgeRAnaPathview(mapped_genes, dir, spe)
}
\arguments{
\item{mapped_genes}{a vector which contain entrez ids}
}
\description{
This function allows you to convert entrez id to genename based on biomart
}
\examples{
edgeRAnaPathview(geneIds)
}
\keyword{convert}
\keyword{id}
