% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgeRAnaReactome.R
\name{edgeRAnaReactome}
\alias{edgeRAnaReactome}
\title{A edgeRAnaReactome Function}
\usage{
edgeRAnaReactome(df, dfCutOff = 0.05, fdr = 0.99, dir, comparisons,
  outprefix, spe)
}
\arguments{
\item{df}{a dataframe with gene differential expression genes}

\item{fdr}{q value cutoff for differentially expressed genes. Numeric value between 0 and 1}

\item{dir}{directory which count files are located}

\item{comparisons}{pairwise comparisons}

\item{outprefix}{the prefix of output file}
}
\description{
This function is to did reactome pathway analysis
}
\examples{
edgeRAnaReactome(pathway, up=FALSE)
edgeRAnaReactome()
}
\keyword{analysis}
\keyword{pathway}
