% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgeRAnaTPM.R
\name{edgeRAnaTPM}
\alias{edgeRAnaTPM}
\title{A edgeRAnaTPM Function}
\usage{
edgeRAnaTPM(files, dir, group, outprefix, spe)
}
\arguments{
\item{files}{a dataframe which contain edgeR pathway analysis results}

\item{dir}{directory which count files are located}

\item{group}{experiment condition, which is a factor}

\item{outprefix}{the prefix of output file}
}
\description{
This function generate the fit model based on edgeR
}
\examples{
edgeRAnaTPM(pathway, up=FALSE)
edgeRAnaTPM()
}
\keyword{analysis}
\keyword{pathway}
