% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgeRAnaWrite.R
\name{edgeRAnaWrite}
\alias{edgeRAnaWrite}
\title{A edgeRAnaWrite Function}
\usage{
edgeRAnaWrite(dfGenes, comparisons, dir, outprefix)
}
\arguments{
\item{dfGenes}{a list of dfGene tables got from edgeRdfGenes}

\item{dir}{directory which count files are located}

\item{outprefix}{the prefix of output file}

\item{names}{a vector stroe the comparison name}

\item{group}{experiment condition, which is a factor}
}
\description{
This function write df Genes in txt files and return these genes
}
\examples{
edgeRAnaWrite()
edgeRAnaWrite()
}
\keyword{analysis}
\keyword{pathway}
