% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgeRdfGenes.R
\name{edgeRdfGenes}
\alias{edgeRdfGenes}
\title{A edgeRdfGenes Function}
\usage{
edgeRdfGenes(fit, coef, pvalue = 1, spe)
}
\arguments{
\item{fit}{fit object generated by edgeRAnaFit}

\item{coef}{pairwise group}
}
\description{
This function allows you to get pathwasy with q-value <0.05
}
\examples{
edgeRdfGenes(pathway, up=FALSE)
edgeRdfGenes()
}
\keyword{analysis}
\keyword{pathway}
