% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble2entrez.R, R/gene2entrez.R
\name{ensemble2entrez}
\alias{ensemble2entrez}
\alias{ensemble2entrez}
\title{A ensemble2entrez Function}
\usage{
ensemble2entrez(genes, spe)

ensemble2entrez(genes, spe)
}
\arguments{
\item{genes}{a vector which contain genel2entrez ids}

\item{spe}{sepecies}

\item{geneIdS}{a vector which contain ensembl2entrez ids}
}
\description{
This function allows you to convert ensemble id to entrez id based on biomart

This function allows you to convert gene id to entrez id based on biomart
}
\examples{
ensemble2entrez(geneIds)
ensemble2entrez()
gene2entrez(geneIds)
gene2entrez()
}
\keyword{convert}
\keyword{id}
