% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble2geneLength.R
\name{ensemble2geneLength}
\alias{ensemble2geneLength}
\title{A ensemble2geneLength Function}
\usage{
ensemble2geneLength(geneIds, spe)
}
\arguments{
\item{geneIdS}{a vector which contain ensembl2entrez ids}
}
\description{
This function allows you to convert ensemble id to entrez id based on biomart
}
\examples{
ensemble2geneLength(geneIds)
ensemble2geneLength()
}
\keyword{convert}
\keyword{id}
