% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entrez2genename.R
\name{entrez2genename}
\alias{entrez2genename}
\title{A entrez2genename Function}
\usage{
entrez2genename(geneIds, spe)
}
\arguments{
\item{geneIdS}{a vector which contain entrez ids}
}
\description{
This function allows you to convert entrez id to genename based on biomart
}
\examples{
entrez2genename(geneIds)
}
\keyword{convert}
\keyword{id}
