% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGeneLengths.R
\name{getGeneLengths}
\alias{getGeneLengths}
\title{A getGeneLengths Function}
\usage{
getGeneLengths(spe)
}
\arguments{
\item{geneIdS}{a vector which contain ensembl2entrez ids}
}
\description{
This function allows you to convert ensemble id to entrez id based on biomart
}
\examples{
getGeneLengths()
getGeneLengths()
}
\keyword{convert}
\keyword{id}
