% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPathway.R
\name{getPathway}
\alias{getPathway}
\title{A getPathway Function}
\usage{
getPathway(pathways, fdr = 0.05, out = T, dir = NA, outprefix = NA)
}
\arguments{
\item{pathways}{a dataframe which contain edgeR pathway analysis results}

\item{fdr}{cutoff of pathway statistic}

\item{out}{whether to write the result to files}

\item{dir}{output directory}

\item{outprefix}{prefix of output files}
}
\description{
This function allows you to get pathwasy with q-value <0.05
}
\examples{
getPathway(pathway, up=FALSE)
getpathway()
}
\keyword{analysis}
\keyword{pathway}
