% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcPlots.R
\name{qcPlots}
\alias{qcPlots}
\title{A qcPlots Function}
\usage{
qcPlots(files, dir, condition, outprefix)
}
\arguments{
\item{files}{the files came from htseq-count (only contain two coloumns)}

\item{dir}{directory which count files are located}

\item{condition}{experiment condition, which is a factor}

\item{outprefiex}{output file name}
}
\description{
This function allows you to draw a pca plot
}
\examples{
qcPlot(files, dir, condtion, outprefix)
qcPlot()
}
\keyword{plot}
\keyword{qc}
