% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roundDataFrame.R
\name{roundDataFrame}
\alias{roundDataFrame}
\title{A roundDataFrame Function}
\usage{
roundDataFrame(x, digits)
}
\arguments{
\item{x}{a dataframe}

\item{digits}{the digits will be round}
}
\description{
This function is to round numeric value in a dataframe
}
\examples{
roundDataFrame()
roundDataFrame()
}
\keyword{math}
