% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble2entrez.R
\name{ensemble2entrez}
\alias{ensemble2entrez}
\title{A ensemble2entrez Function}
\usage{
ensemble2entrez(geneIds, spe)
}
\arguments{
\item{geneIdS}{a vector which contain ensembl2entrez ids}
}
\description{
This function allows you to convert ensemble id to entrez id based on biomart
}
\examples{
ensemble2entrez(geneIds)
ensemble2entrez()
}
\keyword{convert}
\keyword{id}
